Add support for Proxy Protocol to mail SMTP
When Nginx is not the first proxy in a proxy chain it needs to be able to
understand the Proxy Protocol header to retrieve client IP and pass it
upstream in an XCLIENT command. With these changes one can configure
Nginx to do that by adding `proxy_protocol` to a listener description.